+++ lib/carbonite/trigger.ex
@@ -12,12 +12,16 @@ defmodule Carbonite.Trigger do
   @primary_key {:id, :id, autogenerate: true}
   @timestamps_opts [type: :utc_datetime_usec]
 
+  @type mode :: :capture | :ignore
+
   @type t :: %__MODULE__{
           id: non_neg_integer(),
           table_name: String.t(),
           table_prefix: String.t(),
           primary_key_columns: [String.t()],
           excluded_columns: [String.t()],
+          mode: mode(),
+          override_transaction_id: nil | non_neg_integer(),
           inserted_at: DateTime.t(),
           updated_at: DateTime.t()
         }
@@ -27,6 +31,8 @@ defmodule Carbonite.Trigger do
     field(:table_name, :string)
     field(:primary_key_columns, {:array, :string}, default: [])
     field(:excluded_columns, {:array, :string}, default: [])
+    field(:mode, Ecto.Enum, values: [:capture, :ignore])
+    field(:override_transaction_id, :integer)
 
     timestamps()
   end