+++ test/support/test_repo.ex
@@ -6,4 +6,13 @@ defmodule Carbonite.TestRepo do
   use Ecto.Repo,
     otp_app: :carbonite,
     adapter: Ecto.Adapters.Postgres
+
+  import Ecto.Query, only: [from: 2]
+
+  @spec count(module()) :: non_neg_integer()
+  def count(schema) do
+    schema
+    |> from(select: count())
+    |> one!(prefix: Carbonite.default_prefix())
+  end
 end